test(js-themes-toolkit): add assertion functions to troubleshoot gulp failures
This is for:

https://github.com/liferay/liferay-frontend-projects/pull/368

When there is a Gulp failure in the tests of any kind (ie. in the Gulp
code itself, or in test code where we're making assertions) it gets
swallowed and surfaced as:

    Error [ERR_UNHANDLED_ERROR]: Unhandled error. ([object Object])

The `expect()` assertions in the `afterEach` block are particularly bad
because you can't even see which test was running before the post-check
ran. The other trick that can help is to insert a naughty `console.log`
directly in `node_modules/undertaker/lib/helpers/createExtensions.js`
can shed some light.

Anyway, with these assertion functions, you still see the
indescipherable error, but the use of `console.error` at least shows you
a useful stacktrace and hopefully some actually revealing debugging info
anyway.