docs: link to general guidelines from CONTRIBUTING.md
Based on conversation here:

    https://github.com/liferay/liferay-frontend-projects/pull/389#issuecomment-770646774

which was about not finding contributing guidelines for how to format
commit messages.

We could still use some more specific docs about what scopes we'd like
to see when working in the monorepo (eg. using scopes to indicate
project), so I'll put that in the next commit.