feat(npm-tools): add a target to scan JSP files for JS modules to be exposed
This scans for `module=".*"` and `propsTransformer=".*"` and dumps a `npmscripts.config.js`
file or tweaks the existing one.

It's still in experimental phase, but we'll need it sooner or later to ease development.

We may want to change the way it works because touching a `.js` file is not as straightforward as
modifying a `.json` one, so maybe we can find a cleaner way.

Also, we will need to support the equivalent of <aui:script require> once it is ready in liferay-portal.