docs: hoist GPL-2.0.txt license text to top level
The code in "third-party/projects/jquery-form/" is dual-licensed under
the GPL v2.0 and MIT licenses. We're going to be distributing our
version with a tiny patch on top using the MIT because, but it makes
sense to at least have the license text for both licenses available in
the repo.