docs(jquery-form): add CHANGELOG.md for v3.51.0-liferay.1 release
Seeing as this is going to be the first time we release a copy of this
package under the `@liferay` named scope, and there is no CHANGELOG.md
in this project, this one is a little different from a typical update. I
did this:

    cd third-party/projects/jquery-form
    touch CHANGELOG.md
    ../../../projects/npm-tools/packages/changelog-generator/bin/liferay-changelog-generator.js -i

And then manually tweaked the version numbers. I expect the next
release, if there ever is one, will just be a question of running the
generator with no tweaks required, and most likely using the
`--prepatch` switch:

    ../../../projects/npm-tools/packages/changelog-generator/bin/liferay-changelog-generator.js --prepatch