feat(npm-tools): add a helper to extract taglib invocations from JSPs
This is to be able to check taglib invocations so that we can lint the configuration of the exposes
field in webpack.

Note that the function does a bit of pre-work for the lexer, as it scans the JSP for taglib openings
before invoking the lexer, passing it only the tag and its body.

Also note that the lexer is very lazy, in the sense that it only scans the taglib opening to find the
name and attributes and then discards the rest as insignificant data. This is possible because of
the aforementioned pre-work.