feat(npm-scripts): configure if temporary files should be deleted
I'm also fixing the feature, because it wasn't working on successful exit.

The rationale behind this feature is that, for webpack federation, I want to keep temporary files
in the build directory after the build. This is because it contains very useful information for
development/debug purposes and, given that the temporary file is small and written to the build
directory, I think it is better to keep it.