refactor: remove explicit dependency on clayui/css
This came in via 18e9c76f12 but I don't think we actually need this as
a top-level `devDependency`. It will come in anyway transitively with
liferay-frontend-theme-styled and liferay-frontend-theme-unstyled, which
were added in f5da76606f.

Originally looked at this because of this dependabot PR:

    https://github.com/liferay/liferay-frontend-projects/pull/414

I don't think this will stop the bot from reminding us to update our own
internal dependencies (that we aren't even directly using here anyway),
so I might also tell the bot to ignore this dependency.