refactor(npm-scripts): internalize npm-bundler-preset-liferay-dev
Since we don't need two different mechanisms to hold defaults, merge them in a single place.

We put everything inside npm-scripts so that we don't need to release two new versions (of the
preset and npm-scripts) whenever defaults change.

Also, this walks toward the goal of moving imports configuration to liferay-portal project, where it
really belongs.