refactor(npm-scripts): load defaults from root directory
Let projects provide default values for npmsripts.config.js in their root
directory. This is intended to define global remotes and shared primarily, but
can be used for other purposes too.

As a result of these changes, we may not rely on the existence of federation
configuration to autodetect that we should run its process. This is because in
the defaults we will always have federation configuration for every project in
liferay-portal (namely the one defining remotes and shared).

Thus, we introduce a new `mode` setting that can be `disabled`, `compatible` or
`incompatible`, to specify if the federation build must be run and, in that
case, if it should produce an artifact compatible with the previous architecture
(bundler+AMD) or a pure webpack one.

NOTE: This is not a new concept, it was already there, but was not made
explicit, as it was inferred from the federation configuration contents. Making
it explicit is better because it allows to define defaults more easily and
because it can be directly read (because it is explicit).