feat(npm-scripts)!: remove npmscripts presets
We have never used presets and don't think we will need them so we are removing them to
simplify the code.

Note, however, that we used `deepMerge.MODE.OVERWRITE_ARRAYS` to merge the preset and
the user config before, and now we don't do it.

We keep inheriting the preset config from the npmscripts defaults file, but it is better not to use
the overwrite mode any more because otherwise we need to replicate a bunch of code in portal
files whenever we want to add a new remote, shared, dependency, etc.

For now this is enough. Time will tell if we need more sophisticated merging strategies in the
future.