feat(npm-scripts): add types definitions for vfile-message
So I need to add some code to liferay-portal in service of:

    https://issues.liferay.com/browse/LPS-127081
    ("Implement and integrate Global JS Store mechanism in DXP")

and I think it would be very useful to "future proof" it by making sure
it is TypeScript-ready. We probably don't want to have our store just be
un untyped "bag of data", especially if we are going to have to live
with and support this API for many years and it is going to become a
central channel for cross-app communication with DXP. Example libraries
which share this point of view are Undux and Recoil, but they are both
React-specific and we want something generic:

    https://github.com/bcherny/undux
    https://recoiljs.org

So, we can either:

1). Build and ship the API in TypeScript in a way that lends itself to
    strongly typing the data; or:
2). At least _prototype_ the API in TypeScript to show that it supports
    strongly-typed usage, even if we end up stripping the type
    annotations and shipping the thing in JS at the end (ie. leave the
    door open to an eventual conversion to TypeScript).

There are some obstacles to doing either though, because "turning on
TypeScript" in DXP is not going to be as simple as flipping a switch.
The only module where we're using TypeScript currently is doing it in a
sneaky way, via webpack and the ts-loader:

    https://github.com/liferay/liferay-portal/blob/9837c4d81f2de200ad2bcd7acc83578fb25e975c/modules/apps/remote-app/remote-app-client-js/webpack.config.js#L28

To make it possible to _actually_ use TypeScript in a well-integrated
fashion in DXP we're going to have to do a bit more work (that means
making sure all the linting, formatting, testing pathways works,
that our build process doesn't regress in terms of performance, that
cross-module type definitions are easily consumed and available, and so
on...)

This commit is the first in a series of two or three that I'm going to
push into a draft PR just to show what the minimum first steps would be
to get us going and be able to do one of "1" or "2" mentioned above.

There are three main ways to process TypeScript at this time:

1. webpack + ts-loader: this handles both building and type-checking.
2. Babel's `@babel/preset-typescript`, which handles _only_ stripping
   out of TypeScript syntax (eg. annotations); it does not do any actual
   type-checking, so is of limited value.
3. Using the TypeScript compiler itself (`tsc`), which can both
   type-check and emit code.

So given that "2" on its own is not valuable (no type-checking) and
going all-in on "3" is a huge/risky change (ie. from Babel to `tsc`),
I'm going to go with a mixture: we'll run `tsc --noEmit` to do
type-checking only, and then Babel with the preset to do the build while
stripping out TS type annotations.

Testing shows that right now we probably need the type definitions
added in this commit in order for `tsc --noEmit` to work in both
the new code I am adding for the LPS, and the existing code in
`remote-app-client-js`:

    ../../../node_modules/@types/vfile/index.d.ts:11:31 -
    error TS7016: Could not find a declaration file for module
    'vfile-message'. '$ROOT/modules/node_modules/vfile-message/index.js'
    implicitly has an 'any' type.

    Try `npm install @types/vfile-message` if it exists or add
    a new declaration (.d.ts) file containing `declare module
    'vfile-message';`

Specifically, when we tell TS to look in node_modules for type
information, it is going to look at _everything_ in there, and complain
about this module, even though we aren't directly using it.

Now, there is a chance this won't actually work, because we supposedly
_already_ have a copy of the types in DXP via transitive deps
(npm-scripts -> stylelint; see output of `yarn why`):

    => Found "@types/vfile-message@2.0.0"
    info Reasons this module exists
       - "_project_#@liferay#npm-scripts#stylelint#postcss-markdown#remark#unified#@types#vfile" depends on it
       - Hoisted from "_project_#@liferay#npm-scripts#stylelint#postcss-markdown#remark#unified#@types#vfile#@types#vfile-message"
    info Disk size without dependencies: "80KB"
    info Disk size with unique dependencies: "100KB"
    info Disk size with transitive dependencies: "116KB"
    info Number of shared dependencies: 2

And the module _is_ present under `modules/node_modules/@types/`, but I
get the error TS7016 shown earlier until I `yarn add -W --dev` in the
workspace root, after which `yarn why` shows the following and the error
goes away:

    => Found "@types/vfile-message@2.0.0"
    info Has been hoisted to "@types/vfile-message"
    info Reasons this module exists
       - "workspace-aggregator-b1199f56-1776-47db-b508-813de5c8a825" depends on it
       - Specified in "devDependencies"
       - Hoisted from "_project_#@types#vfile-message"
       - Hoisted from "_project_#@liferay#npm-scripts#stylelint#postcss-markdown#remark#unified#@types#vfile#@types#vfile-message"
    info Disk size without dependencies: "12KB"
    info Disk size with unique dependencies: "76KB"
    info Disk size with transitive dependencies: "108KB"
    info Number of shared dependencies: 3

For whatever reason, when the error is present, Yarn is installing a
nested copy of vfile-message (v2.0.2) inside `@types/vfile-message`
because elsewhere via a lengthy dependency chain we're getting v1.1.1
(via npm-scripts -> stylelint -> postcss-markdown -> remark -> unified
-> vfile, which asks for "^0.0.0"):

    => Found "vfile-message@1.1.1"
    info Reasons this module exists
       - "_project_#@liferay#npm-scripts#stylelint#postcss-markdown#remark#unified#vfile" depends on it
       - Hoisted from "_project_#@liferay#npm-scripts#stylelint#postcss-markdown#remark#unified#vfile#vfile-message"
    info Disk size without dependencies: "20KB"
    info Disk size with unique dependencies: "36KB"
    info Disk size with transitive dependencies: "36KB"
    info Number of shared dependencies: 1
    => Found "@types/vfile-message#vfile-message@2.0.2"
    info This module exists because "_project_#@liferay#npm-scripts#stylelint#postcss-markdown#remark#unified#@types#vfile#@types#vfile-message" depends on it.
    info Disk size without dependencies: "28KB"
    info Disk size with unique dependencies: "60KB"
    info Disk size with transitive dependencies: "60KB"
    info Number of shared dependencies: 2

Adding the dependency explicitly in DXP itself, the nested copy goes
away, and along with it, the error. We'll see if adding it here does
the trick too.

The grand irony of all of this is that if we were on a newer version of
`vfile-message`, we probably wouldn't need any of this anyway, because
the `@types` package is actually deprecated and advises people that
`vfile-message` bundles its own types. Sadly, it seems that explaining
all this is still less work and less risk than attempting a big
stylelint update right now.