feat(npm-scripts): tell Babel to process TS files
Without this, Babel won't look at ".ts" files and you'll get

    Successfully compiled 0 files with Babel.

Because the default list documented here:

    https://babeljs.io/docs/en/babel-core#default_extensions

Contains only:

    ".js", ".jsx", ".es6", ".es", ".mjs", ".cjs"

It would obviously be better to configure this via a config file because
we don't have a mechanism for passing arbitrary arguments through to
`babel` (we don't want a wild-west situation where every team invokes
Babel in their own special way) but as it says on the linked page:

> Extending this list isn't possible, however `@babel/cli` does provide
> ways to support other extensions with `--extensions`.

There is an open ticket for improving this in the Babel repo:

    https://github.com/babel/babel/issues/8652

but it is from September 2018 and there hasn't been much activity, so it
may be a while before this is improved.

So, we add ".ts" (needed imminently) and ".tsx" (for future-proofness).

Tested on existing module that contains TypeScript in DXP
(remote-app-client-js) and this seems to work fine. Note that that
module doesn't care what Babel does because it does the real build with
webpack, but this will come in handy for our work on:

    https://issues.liferay.com/browse/LPS-127081

(where we want to do a more normal/integrated build, as explained in the
parent commit).