feat(npm-scripts): check TypeScript types during build
As explained in the grandparent commit, we can tell Babel to strip type
annotations for us, but it won't actually check the types. We have to do
that explicitly with `tsc` ourselves.

In the case of modules like `remote-app-client-js` that means we
actually redundantly check the types twice (once with tsc and once with
webpack's ts-loader), but that project is tiny so the duplicate work
isn't noticeable, and we don't expect to make more modules like this in
the future (although as we move towards the new "federated" build style
we will be leaning less heavily on Babel and the balance may tip in the
opposite direction).