fix(npm-scripts): change the way we merge check and fix scripts
We need to be able to overwrite these two arrays (instead of merge) to make the tool useful.

This has been discovered while running the unit tests, but apart from that, we have configuration
in liferay-portal that makes this a must. In particular, frontend-js-metal-web,
frontend-js-node-shims, and frontend-taglib-chart empty these two fields.