feat(npm-scripts): new temporary option to turn off bundler completely
I'm adding an option named `bundling` which can be set to `false` to turn off the bundler alone.

The rationale for this need is that while we are migrating the portal to webpack, some modules,
after migration, will have the `mode` federation option set to `disabled` (this is because they
won't need webpack at all; usually because they are very low level or very old bundles containing
raw JS code).

While we are migration, setting `mode` to `disabled` causes the bundler to kick in, which is
something we don't want for these modules either. Because of that, I'm adding this new option
to temporarily mark modules which are completely migrated.

Once we remove the bundler from npm-scripts we may get rid of this option.