feat(npm-scripts): activate TypeScript preset by default
While rebasing my WIP liferay-portal PR:

    https://github.com/wincent/liferay-portal/pull/373

I had to resolve conflicts because we've had a concurrent update to
`@liferay/npm-scripts` in the meantime. In order to continue developing
that PR, I am going to need a release of the scripts.

Given that the draft PR I shared here:

    https://github.com/liferay/liferay-frontend-projects/pull/425

got some "lg2m" comments, I may as well merge it so that it will wind up
in the next scripts release, and I won't need to keep releasing new
previews just to keep up.

This commit is the one change that I think will be needed in order to
make this changes "safe" for liferay-portal. Without this preset, the
build will choke in layout-content-page-editor-web because it contains
".d.ts" files that Babel would normally ignore, but will now try to
parse; activating the preset prevents the build from choking on the
syntax. The preset is pretty benign; all it does is teach Babel to grok
TypeScript syntax and strip it out during the build (no type-checking
actually occurs during the Babel run).

Test plan: I will cut one more test release after this PR is pushed, and
if an `ant all` works with that, then I'd say we'll be safe to merge it.