feat(eslint-config)!: turn off no-for-of-loops rule by default
Now that we no longer support IE on the DXP master branch, we can turn
this rule off by default, and then opt explicitly back into it on the
maintenance branches (eg. 7.3.x and older), or in products which go to
the maintenance branches (eg. alloy-editor etc).

The truth is that even in places where we need to support IE, this rule
was always very conservative; we polyfill/transpile in IE, and the only
reason that that couldn't be relied upon to work swimmingly is that
Babel used `Symbol.iterator` in its output and we weren't 100% confident
in our polyfill coverage:

https://github.com/liferay/liferay-frontend-projects/issues/354

In any case, I'll be following up with some pulls to other
projects/branches to make turn the rule back on where it is still
needed.

One final note: I am turning the rule `'off'` rather than just removing
it so as to reflect that we still do use this rule in some places, it's
just that we don't have it on by default.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/405