fix(eslint-config): update monkey patching to work with newer ESLint
While following up on:

    https://github.com/liferay/liferay-frontend-projects/pull/429

I went to update alloy-editor's lint set-up (which is very old) and
noticed that ESLint has been doing some refactoring that means that
newer versions don't work properly with your monkey-patching technique,
and need to be updated again.

Now, the monkey patching has _always_ sucked, and we know how to
restructure our code so that we won't need any more, but it will require
quite a bit of work; see:

    https://github.com/liferay/liferay-frontend-projects/issues/9

(In short, creating a plugin package and using that instead of the
existing config package, because configs cannot contain plugins but
plugins can contain configs.)

Anyway, I don't want to do that now because it is probably a full day's
work and right now I just want to get unblocked, so this commit goes
ahead and updates the monkey patch.