feat(eslint-config): teach no-duplicate-imports rule about type imports
It's legit to import from the same module twice, as long as exactly one
of the imports is a type-only import:

    import Thing from './Thing';
    import type {ThingT} from './Thing';

In the old days, Flow had this syntax and TypeScript didn't. You would
instead have to import the type as though it were a value like this, and
trust the TypeScript compiler to figure out that you only wanted it for
the type definition:

    import Thing, {ThingT} from './Thing';

This made it tricky to do things like circular type-only imports (ie.
modules A and B wanting to share type definitions with each other). But
TypeScript has support for this now and the way we're currently using
this it in liferay-portal, we don't even have a choice really, because
we're using Babel to strip out type information (Babel knows nothing
about the types unless they are flagged syntactically, so if it sees a
mixed type/non-type import it doesn't know that it has to strip out part
of it).