fix(npm-scripts): change Angular adaptation to make it work with Angular 10 and 11
Starting at Angular 10, the ng-cli does not longer generate JS files for ES5 and ES2015, but just
one flavor. To account for that, we create a loader that discards ES2015 files from Angular <=9
and removes the -es5 particle from the name.

That way everything is treated as if it was coming from Angular 10.

Note however, that this is a dirty hack that cannot be made cleaner because we would have to
change the way bundler loaders work.