fix: don't treat ourselves as a valid import
This is needed since we allowed to inherit the base configuration from the liferay-portal project in
npm-scripts.

This is because now we don't have a way to disable the inherited imports other than setting them
to false one by one. But that makes it very difficult to configure projects, because if we add a new
import to the base config, we need to check all other projects one by one.

With this change, we can use the same base config in all projects without making the build fail
because a project excludes itself from the JAR because the bundler thought it was imported
(rather than the project being built).