chore(monorepo): temporarily fix lockfile failure in CI
CI is failing on master (cfb2c217c449f717e473) right now:

    https://github.com/liferay/liferay-frontend-projects/runs/2140506487

because we just prepared the v23.0.0 release of
`@liferay/eslint-config`, but `@liferay/npm-scripts` still depends on
v22.0.0.

Now, @kresimir-coko will fix that tomorrow when he completes the release
and updates the dependency, so for now I am just freshening the lockfile
to keep CI green. (I could also fix this by updating the dependency
myself, but that would require me to finish the release, and I don't
want to do that.)